GraphServer: Remove reliance on CloudSpanner's TypeCode
The google.cloud.spanner.TypeCode was used to validate supported
types for properties. However, the type information is always
received as a string from the JS client. It is only internally converted
to a TypeCode for verification.

Comparing against the TypeCode does not make the validation any more
robust than just confirming if the strings themselves are valid type
strings.

Removing reliance on TypeCode also makes GraphServer database agnostic.