GraphServer: Remove reliance on CloudSpanner's TypeCode (#48)
* Standardize return type for execute_query() functions

- Also removes cloud-spanner specific fields from the return type.
  Specifically `StructType.Field` is removed from the return type.
  Removing this tightly coupled logic is required to allow new DB
  implementations.

* Abstract away "database" implementations and remove strong coupling of DB implementation with APIs

1. Abstracts SpannerDatabase with clear APIs
2. Introduces CloudSpannerDatabase as an implementation of SpannerDatabase
3. Removes further tight coupling with the cloud spanner client by adding a
   SpannerFieldInfo dataclass to replace usage of StructType.Field

* Introduce exec_env.py to maintain global state + minor bug fix

1. The global database_instances is moved to exec_env.py to avoid
   circular imports.

2. SpannerFiledInfo.typename populated with the correct name now

3. Remove all cloud spanner refs from database.py

* GraphServer: Remove reliance on CloudSpanner's TypeCode

The google.cloud.spanner.TypeCode was used to validate supported
types for properties. However, the type information is always
received as a string from the JS client. It is only internally converted
to a TypeCode for verification.

Comparing against the TypeCode does not make the validation any more
robust than just confirming if the strings themselves are valid type
strings.

Removing reliance on TypeCode also makes GraphServer database agnostic.

* SpannerQueryResult: Change 'error' to 'err'

All consumers of SpannerQueryResult expect `err` instead of `error`.
Without this fix, failures are silent. With this, they're displayed
within the notebook.

Thanks to @cqian23 for pointing this out

---------

Co-authored-by: Sailesh Mukil <mukil.sailesh@gmail.com>
Co-authored-by: Sailesh Mukil Gangatharan <saileshmukil@google.com>